requirements.txt
pdfplumber
faiss-cpu
sentence-transformers
rapidfuzz
streamlit
openai
peft